requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.0
djangorestframework==3.13.1
djongo==1.3.5
drf-writable-nested==0.7.0
pymongo==3.12.3
python-decouple==3.6
pytz==2022.2.1
sqlparse==0.2.4
